Camera/CameraNormal: ctrlNormalOrTower 95.18% -> 95.41% (retain full yaw)

+++ src/Camera/CameraNormal.cpp
@@ -166,7 +166,7 @@ void CPolarSubCamera::ctrlNormalOrTowerCamera_()
 						                      sf);
 					}
 				} else {
-					s16 sVar9 = *gpMarioAngleY - 0x8000;
+					int sVar9 = *gpMarioAngleY - 0x8000;
 					f32 f29;
 					f32 f30;
